# Conflicts:
#	alo7_inf/ddl/v_a10_album.sql
#	alo7_inf/ddl/v_a11_video.sql
#	alo7_inf/ddl/v_a13_work.sql
#	alo7_inf/ddl/v_album_video_count.sql
#	alo7_inf/ddl/v_album_work_count.sql
#	alo7_inf/src/main/java/cn/com/alo7/inf/common/utils/PageUtils.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/controller/AlbumController.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/controller/BaseController.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/entity/AlbumFullView.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/entity/WorkFullView.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/repository/AlbumFullViewRepository.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/repository/WorkFullViewRepository.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/service/IAlbumViewService.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/service/IWorkService.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/service/IWorkViewService.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/service/impl/AlbumViewServiceImpl.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/service/impl/WorkViewServiceImpl.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/vo/RelationshipListVo.java
#	alo7_inf/src/main/java/cn/com/alo7/inf/vo/RelationshipVo.java
#	alo7_inf/src/main/resources/application-local.yml
#	alo7_inf/target/classes/META-INF/maven/cn.com.broadtext/alo7_inf/pom.properties
#	alo7_inf/target/classes/cn/com/alo7/inf/controller/AlbumController.class

+++ alo7_inf/src/main/java/cn/com/alo7/inf/entity/AlbumView.java
@@ -62,7 +62,7 @@ public class AlbumView extends BaseEntity{
 	private String description;
 	
 	private String status;
-	private Integer position;
+	private Long position;
 	public Long getId() {
 		return id;
 	}
@@ -123,10 +123,10 @@ public class AlbumView extends BaseEntity{
 	public void setStatus(String status) {
 		this.status = status;
 	}
-	public Integer getPosition() {
+	public Long getPosition() {
 		return position;
 	}
-	public void setPosition(Integer position) {
+	public void setPosition(Long position) {
 		this.position = position;
 	}
 }
