+++ alo7_inf/src/main/java/cn/com/alo7/inf/vo/AlbumVo.java
@@ -18,7 +18,7 @@ public class AlbumVo extends BaseEntity{
 	private String backgroundImgUrl;
 	private String description;
 	private String status;
-	private Integer position;
+	private Long position;
 	public Long getId() {
 		return id;
 	}
@@ -79,10 +79,10 @@ public class AlbumVo extends BaseEntity{
 	public void setStatus(String status) {
 		this.status = status;
 	}
-	public Integer getPosition() {
+	public Long getPosition() {
 		return position;
 	}
-	public void setPosition(Integer position) {
+	public void setPosition(Long position) {
 		this.position = position;
 	}
 }
